forecasting-product/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forecasting-product"
version = "0.2.0"
description = "Modular time series forecasting product for retail/supply-chain S&OP"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "polars>=0.20.0",
    "pyyaml>=5.4.0",
    "statsforecast>=1.6.0",
    "mlforecast>=0.12.0",
    "lightgbm>=3.3.0",
    "xgboost>=1.5.0",
    "scikit-learn>=1.0.0",
    "duckdb>=0.9.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
]
auth = [
    "pyjwt>=2.0.0",
    "bcrypt>=4.0.0",
]
sku-mapping = [
    "rapidfuzz>=3.0.0",
]
explainability = [
    "shap>=0.42.0",
]
holidays = [
    "holidays>=0.40",
]
spark = [
    "pyspark>=3.4.0",
    "delta-spark>=2.4.0",
]
azure = [
    "azure-identity>=1.15.0",
    "azure-storage-file-datalake>=12.14.0",
]
all = [
    "forecasting-product[dev,auth,sku-mapping,explainability,holidays,spark,azure]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # line length handled by formatter
